KPM

KPM.fermiFunction

KPM.fermiFunctions

KPM.kpm_1d

KPM.kpm_1d!

KPM.kpm_2d

KPM.kpm_2d!

KPM.sigmaMatricesDot

KPM.fermiFunction — Method

fermiFunction(E, E_f, beta)

calculate Fermi-Dirac function at energy E, Fermi energy μ and temperature β =1/T. Input and output all Float64. Infinite β only allowed when accessing fermi energy through fermiFunctions(). [For performance reason for now. TODO: allow β=Inf here withouth perf. reduction. ]

Allow sloppy use of type as long as convertion is available, if using keyword arguments.

KPM.fermiFunctions — Method

fermiFunctions(E_f::Float64, beta::Float64)

returns a fermi function with given E_f and beta.

Allow sloppy use of type as long as convertion is available, if using keyword arguments.

KPM.kpm_1d! — Method

kpm_1d!(H, NC, NR, NH, mu; psi_in_l, psi_in_r, verbose)

Calculate the moments μ defined in KPM. Output is saved in mu.

Fields

H – Hamiltonian. A matrix or sparse matrix

NC – Integer. the cut off dimension

NR – Integer. number of random vectors used for KPM evaluation

NH – Integer. the size of hamiltonian

mu – Array. Output

psi_in_l – Array. Input array on the left side. A ket as psi_in_r.

Hence, when calculating trace like in DOS, psi_in_l == psi_in_r.

psi_in_r – Array. Input array on the right side. A ket. Hence for

example, when calculating trace like in DOS, psi_in_l == psi_in_r.

force_norm – Boolean, Optional. Apply normalization.

KPM.kpm_1d — Method

kpm_1d(H, NC, NR, NH; psi_in, psi_in_l, psi_in_r, force_norm, verbose)

Calculate the moments μ defined in KPM.

Fields

H – Hamiltonian. A matrix or sparse matrix

NC – Integer. the cut off dimension

NR – Integer. number of random vectors used for KPM evaluation

NH – Integer. the size of hamiltonian

psi_in – Optional. Allow setting random vector manually.

force_norm – Boolean, Optional. Apply normalization.

verbose – Integer. Default is 0. Enables progress bar if set verbose=1.

KPM.kpm_2d! — Method

kpm_2d!(H, Jα, Jβ, NC, NR, NH, μ; arr_size, psi_in_l, psi_in_r, ψ0r, Jψ0r, JTnHJψr, ψall_r, ψ0l, ψall_l, ψw)

In place KPM2D. This is also the main building block for KPM_2D. This method only provide NR=1.

Calculates ψ0l * Tm(H) * Jβ * Tn(H) * Jα * ψ0r. When ψ0r and ψ0l are chosen to be random and identical, the output approximates tr(Tm(H) Jβ Tn(H) Jα). The accuracy is ~ O(1/sqrt(NR * NH)) with NR repetitions. NC controls the energy resolution of the result.

Output: nothing. Result is saved on μ.

ARGS

`H: Hamiltonian. A sparse 2D array.

`Jα: Current operator. A sparse 2D array.

`Jβ: Current operator. A sparse 2D array.

`NC: Integer. KPM cutoff order.

`NR: Integer. Number of random vectors.

`NH: Integer. Dimension of H, Jα and Jβ

`μ: 2D Array of dimension (NC, NC). Results will be updated here. Any data

will be overwritten.

KWARGS

`arr_size: The buffer array size. Minimum is 3. Determines the number of

left states to be kept in memory for each loop of right states. The time complexity is reduced from O(NNC^2) to O(NNCarr_size) while space complexity is increased from O(NNC) to O(NNCarr_size).

`psiinl: Passes value to ψ0l. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psiinl should be normalized.

`psiinr: Passes value to ψ0r. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psiinr should be normalized.

working spaces KWARGS: The following keyword args are simply providing working place arrays to avoid repetitive allocation and GC. They are automatically created if not set. However, when using KPM_2D! for many times, it is beneficial to reuse those arrays. CONVENTION: args with 'ψ' are all working space arr.

ψ0r=maybe_on_device_zeros(NH, NR)

Jψ0r=maybe_on_device_zeros(NH, NR)

JTnHJψr=maybe_on_device_zeros(NH, NR)

ψall_r=maybe_on_device_zeros(3, NH, NR)

ψ0l=maybe_on_device_zeros(NH, NR)

ψall_l=maybe_on_device_zeros(arr_size, NH, NR)

ψw=maybe_on_device_zeros(NH, NR)

KPM.kpm_2d — Method

kpm_2d(H, Ja, Jb, NC, NR, NH; psi_in, psi_in_l, psi_in_r, kwargs...)

Calculate moments for 2D KPM.

Calculates ψ0l * Tm(H) * Jβ * Tn(H) * Jα * ψ0r. When ψ0r and ψ0l are chosen to be random and identical, the output approximates tr(Tm(H) Jβ Tn(H) Jα). The accuracy is ~ O(1/sqrt(NR * NH)). NC controls the energy resolution of the result.

Output: μ, a 2D array in ComplexF64. μ[n, m] is the momentum for 2D KPM.

ARGS

H Hamiltonian. A sparse 2D array.

Jα Current operator. A sparse 2D array.

Jβ Current operator. A sparse 2D array.

NC Integer. KPM cutoff order.

NR Integer. Number of random vectors to choose from. When skipped, understood as NR=1.

NH Integer. Dimension of H, Jα and Jβ

KWARGS

psi_in_l

Passes value to ψ0l. The array is not updated. Size should be (NH, NR) (preferred) or (NR, NH) if set.

psi_in_r

Passes value to ψ0r. The array is not updated. Size should be (NH, NR) (preferred) or (NR, NH) if set.

psi_in

Cannot be used together with psiinl and psiinr. Sets psiinl=psiinr=psi_in if set.

kwargs

other kwargs in KPM_2D!

KPM.sigmaMatricesDot — Method

wi dot σi i correspond to x,y,z and last one is identity examples: smd3 = sigmaMatricesDot(3) smd3([1 0 0 0]) -> 2×2 Array{ComplexF64,2}: 0.0+0.0im 1.0+0.0im 1.0+0.0im 0.0+0.0im smd3([0 1 0 0]) -> 2×2 Array{ComplexF64,2}: 0.0+0.0im 0.0-1.0im 0.0+1.0im 0.0+0.0im smd3([1 1 1 1]) -> 2×2 Array{ComplexF64,2}: 2.0+0.0im 1.0-1.0im 1.0+1.0im 0.0+0.0im