KPM

KPM.fermiFunction

KPM.fermiFunctions

KPM.kpm_1d

KPM.kpm_1d!

KPM.kpm_2d

KPM.kpm_2d!

KPM.fermiFunction — Method

fermiFunction(E, E_f, beta)

calculate Fermi-Dirac function at energy E, Fermi energy μ and temperature β =1/T. Input and output all Float64. Infinite β only allowed when accessing fermi energy through fermiFunctions(). [For performance reason for now. TODO: allow β=Inf here withouth perf. reduction. ]

Allow sloppy use of type as long as convertion is available, if using keyword arguments.

KPM.fermiFunctions — Method

fermiFunctions(E_f::Float64, beta::Float64)

returns a fermi function with given E_f and beta.

Allow sloppy use of type as long as convertion is available, if using keyword arguments.

KPM.kpm_1d! — Method

kpm_1d!(H::Any, NC::Int64, NR::Int64, NH::Int64, mu::Any, psi_in::Any; verbose, α_all, μ_all)

Calculate the moments μ defined in KPM. Output is saved in mu.

H – Hamiltonian. A matrix or sparse matrix

NC – Integer. the cut off dimension

NR – Integer. number of random vectors used for KPM evaluation

NH – Integer. the size of hamiltonian

mu – Array. Output

psi_in – Array (optional). Input array on the right side. A ket.

KPM.kpm_1d — Method

kpm_1d(H::Any, NC::Int64, NR::Int64, NH::Int64; psi_in, psi_in_l, psi_in_r, force_norm, verbose)

Calculate the moments μ defined in KPM.

H – Hamiltonian. A matrix or sparse matrix

NC – Integer. the cut off dimension

NR – Integer. number of random vectors used for KPM evaluation

NH – Integer. the size of hamiltonian

psi_in – Optional. Allow setting random vector manually.

force_norm – Boolean, Optional. Apply normalization.

verbose – Integer. Default is 0. Enables progress bar if set verbose=1.

KPM.kpm_2d! — Method

kpm_2d!(H::Any, Jα::Any, Jβ::Any, NC::Int64, NR::Int64, NH::Int64, μ::Any, psi_in_l::Any, psi_in_r::Any; arr_size, verbose, ψ0r, Jψ0r, JTnHJψr, ψall_r, ψ0l, ψall_l, ψw)

In place KPM2D. This is also the main building block for KPM_2D. This method only provide NR=1.

Calculates ψ0l * Tm(H) * Jβ * Tn(H) * Jα * ψ0r. When ψ0r and ψ0l are chosen to be random and identical, the output approximates tr(Tm(H) Jβ Tn(H) Jα). The accuracy is $\sim O(1/sqrt(NR * NH))$ with NR repetitions. NC controls the energy resolution of the result.

Output: nothing. Result is saved on μ.

ARGS

H : Hamiltonian. A sparse 2D array.

Jα : Current operator. A sparse 2D array.

Jβ : Current operator. A sparse 2D array.

NC : Integer. KPM cutoff order.

NR : Integer. Number of random vectors.

NH : Integer. Dimension of H, Jα and Jβ

μ : 2D Array of dimension (NC, NC). Results will be updated here. Any data

will be overwritten.

psi_in : Set psi_in_l and psi_in_r. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psi_in should be normalized.

psi_in_l : Passes value to ψ0l. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psi_in_l should be normalized. psi_in_l is given as column vector of ket $|ψl> = <ψl|^\dagger$

psi_in_r : Passes value to ψ0r. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psi_in_r should be normalized. psi_in_r is given as column vector of ket $|ψr>$.

KWARGS

arr_size : The buffer array size. Minimum is 3. Determines the number of

left states to be kept in memory for each loop of right states. The time complexity is reduced from $O(N\times NC^2)$ to $O(N\times NC\times arr\_size)$ while space complexity is increased from $O(N\times NC)$ to $O(N\times NC\times arr\_size)$.

working spaces KWARGS: The following keyword args are simply providing working place arrays to avoid repetitive allocation and GC. They are automatically created if not set. However, when using KPM_2D! for many times, it is beneficial to reuse those arrays. CONVENTION: args with ψ are all working space arr.

ψ0r=maybe_on_device_zeros(NH, NR)

Jψ0r=maybe_on_device_zeros(NH, NR)

JTnHJψr=maybe_on_device_zeros(NH, NR)

ψall_r=maybe_on_device_zeros(3, NH, NR)

ψ0l=maybe_on_device_zeros(NH, NR)

ψall_l=maybe_on_device_zeros(arr_size, NH, NR)

ψw=maybe_on_device_zeros(NH, NR)

KPM.kpm_2d — Method

kpm_2d(H::Any, Ja::Any, Jb::Any, NC::Int64, NR::Int64, NH::Int64; psi_in, psi_in_l, psi_in_r, kwargs...)

Calculate moments for 2D KPM.

Calculates ψ0l * Tm(H) * Jβ * Tn(H) * Jα * ψ0r. When ψ0r and ψ0l are chosen to be random and identical, the output approximates tr(Tm(H) Jβ Tn(H) Jα). The accuracy is ~ O(1/sqrt(NR * NH)). NC controls the energy resolution of the result.

Output: μ, a 2D array in ComplexF64. μ[n, m] is the momentum for 2D KPM.

ARGS

H Hamiltonian. A sparse 2D array.

Jα Current operator. A sparse 2D array.

Jβ Current operator. A sparse 2D array.

NC Integer. KPM cutoff order.

NR Integer. Number of random vectors to choose from. When skipped, understood as NR=1.

NH Integer. Dimension of H, Jα and Jβ

KWARGS

psi_in_l

Passes value to ψ0l. The array is not updated. Size should be (NH, NR) (preferred) or (NR, NH) if set.

psi_in_r

Passes value to ψ0r. The array is not updated. Size should be (NH, NR) (preferred) or (NR, NH) if set.

psi_in

Cannot be used together with psiinl and psiinr. Sets psiinl=psiinr=psi_in if set.

kwargs

other kwargs in KPM_2D!