KPM

KPM.JacksonKernel

KPM.JacksonKernel

KPM.LorentzKernels

KPM.LorentzKernels

KPM.broadcast_dot_1d_1d!

KPM.broadcast_dot_2d_1d!

KPM.broadcast_dot_reduce_avg_2d_1d!

KPM.chebyshev_iter

KPM.dos

KPM.dos0

KPM.fermiFunction

KPM.fermiFunctions

KPM.isNotBoundary

KPM.kpm_1d

KPM.kpm_1d!

KPM.kpm_2d

KPM.kpm_2d!

KPM.normalizeH

KPM.normalize_by_col

KPM.wrapAdd

Moment calculation

The first step in KPM is calculating moments using Hamiltonians (and current operators for conductivity etc.) Functions with ! are more efficient in-place version; those without ! are convenient methods based on the in-place methods.

KPM.kpm_1d — Function

kpm_1d(H, NC, NR; kwargs...)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:178.

kpm_1d(H, NC, NR, NH; psi_in, psi_in_l, psi_in_r, force_norm, verbose, avg_output)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:179.

The simple version of 1D KPM that returns the moment.

H – Hamiltonian. A matrix or sparse matrix

NC – Integer. the cut off dimension

NR – Integer. number of random vectors used for KPM evaluation

NH – Integer. the size of hamiltonian

psi_in – Optional. Allow setting random vector manually.

force_norm – Boolean, Optional. Apply normalization.

verbose – Integer. Default is 0. Enables progress bar if set verbose=1.

avg_output – Boolean. Default is true. Whether to output averaged μ (hence size NC) or separate μs (size NR x NC).

KPM.kpm_1d! — Function

kpm_1d!(H, NC, NR, NH, mu_all, psi_in; verbose, α_all)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:219.

kpm_1d!(H, NC, NR, NH, mu; kwargs...)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:278.

kpm_1d!(H, NC, NR, NH, mu, psi_in_l, psi_in_r; kwargs...)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:287.

The in-place version of 1D KPM. Calculate the moments μ defined in KPM. Output is saved in mu.

H – Hamiltonian. A matrix or sparse matrix.

NC – Integer. the cut off dimension.

NR – Integer. number of random vectors used for KPM evaluation.

NH – Integer. the size of hamiltonian.

mu_all – Array. Output for each random vector. Size (NR, NC).

psi_in – Array (optional). Input array on the right side. A ket.

KPM.kpm_2d — Function

kpm_2d(H, Ja, Jb, NC, NR, NH; psi_in, psi_in_l, psi_in_r, kwargs...)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:299.

The simple version of 2D KPM that returns the moment. Calculate moments for 2D KPM.

Calculates ψ0l * Tm(H) * Jβ * Tn(H) * Jα * ψ0r. When ψ0r and ψ0l are chosen to be random and identical, the output approximates tr(Tm(H) Jβ Tn(H) Jα). The accuracy is ~ O(1/sqrt(NR * NH)). NC controls the energy resolution of the result.

Output: μ, a 2D array in ComplexF64. μ[n, m] is the momentum for 2D KPM.

ARGS

H

Hamiltonian. A sparse 2D array.

Jα

Current operator. A sparse 2D array.

Jβ

Current operator. A sparse 2D array.

NC

Integer. KPM cutoff order.

NR

Integer. Number of random vectors to choose from. When skipped, understood as NR=1.

NH

Integer. Dimension of H, Jα and Jβ

KWARGS

psi_in_l

Passes value to ψ0l. The array is not updated. Size should be (NH, NR) (preferred) or (NR, NH) if set.

psi_in_r

Passes value to ψ0r. The array is not updated. Size should be (NH, NR) (preferred) or (NR, NH) if set.

psi_in

Cannot be used together with psiinl and psiinr. Sets psiinl=psiinr=psi_in if set.

kwargs

other kwargs in KPM_2D!

KPM.kpm_2d! — Function

kpm_2d!(H, Jα, Jβ, NC, NR, NH, μ, psi_in_l, psi_in_r; arr_size, verbose, ψ0r, Jψ0r, JTnHJψr, ψall_r, ψ0l, ψall_l, ψw)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:310.

kpm_2d!(H, Jα, Jβ, NC, NR, NH, μ, psi_in; kwargs...)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:402.

kpm_2d!(H, Jα, Jβ, NC, NR, NH, μ; kwargs...)

defined at /home/runner/work/KPM.jl/KPM.jl/src/moment.jl:412.

In place KPM2D. This is also the main building block for KPM_2D. This method only provide NR=1.

Calculates ψ0l * Tm(H) * Jβ * Tn(H) * Jα * ψ0r. When ψ0r and ψ0l are chosen to be random and identical, the output approximates tr(Tm(H) Jβ Tn(H) Jα). The accuracy is $\sim O(1/sqrt(NR * NH))$ with NR repetitions. NC controls the energy resolution of the result.

Output: nothing. Result is saved on μ.

ARGS

H : Hamiltonian. A sparse 2D array.

Jα : Current operator. A sparse 2D array.

Jβ : Current operator. A sparse 2D array.

NC : Integer. KPM cutoff order.

NR : Integer. Number of random vectors.

NH : Integer. Dimension of H, Jα and Jβ

μ : 2D Array of dimension (NC, NC). Results will be updated here. Any data

will be overwritten.

psi_in : Set psi_in_l and psi_in_r. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psi_in should be normalized.

psi_in_l : Passes value to ψ0l. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psi_in_l should be normalized. psi_in_l is given as column vector of ket $|ψl> = <ψl|^\dagger$

psi_in_r : Passes value to ψ0r. Size is (NH, NR). The array is not updated.

Whether the input is normalized or not, it is assumed to be intended. Usually psi_in_r should be normalized. psi_in_r is given as column vector of ket $|ψr>$.

KWARGS

arr_size : The buffer array size. Minimum is 3. Determines the number of

left states to be kept in memory for each loop of right states. The time complexity is reduced from $O(N\times NC^2)$ to $O(N\times NC\times arr\_size)$ while space complexity is increased from $O(N\times NC)$ to $O(N\times NC\times arr\_size)$.

working spaces KWARGS: The following keyword args are simply providing working place arrays to avoid repetitive allocation and GC. They are automatically created if not set. However, when using KPM_2D! for many times, it is beneficial to reuse those arrays. CONVENTION: args with ψ are all working space arr.

ψ0r=maybe_on_device_zeros(NH, NR)

Jψ0r=maybe_on_device_zeros(NH, NR)

JTnHJψr=maybe_on_device_zeros(NH, NR)

ψall_r=maybe_on_device_zeros(3, NH, NR)

ψ0l=maybe_on_device_zeros(NH, NR)

ψall_l=maybe_on_device_zeros(arr_size, NH, NR)

ψw=maybe_on_device_zeros(NH, NR)

applications

DOS

To calculate density of state (DOS), first calculating moment first using kpm_1d or kpm_1d! with default (random) input vectors. Then use the output (moment mu) to calculate density of state. There is also an option to pass Hamiltonian directly to dos, which does the moment calculation automatically.

KPM.dos — Function

dos(H; NC, NR, E_grid, N_tilde, E_range, kernel, fix_normalization)

defined at /home/runner/work/KPM.jl/KPM.jl/src/applications/dos.jl:31.

dos(μ, H_rescale_factor; E_grid, N_tilde, E_range, NC, kernel)

defined at /home/runner/work/KPM.jl/KPM.jl/src/applications/dos.jl:48.

Calculate DOS for a fermi energy grid spanning E_range with N_tilde total points. If E_range is not set, automatically set it to be sightly smaller than full size. Otherwise an explicit array of E_grid can be passed in. Don't do both.

Either a) pass in a 1d array as moment and as normalization factor; or b) pass in a Hamiltonian that is rescaled with an optional keyword rescale_factor that default to 1.

H_rescale_factor is the normalization of H. Needed when μ is passed.

NR random vectors. Needed when H is passed

Kernels

Kernels are functions defined as

kernel(n::Int64, N::Int64) -> Float64

such that when n==0, returns 1; when n==N-1, returns small number close to 0.

We implement JacksonKernel and LorentzKernels in the package. Jackson Kernel is the default kernel for most application.

KPM.JacksonKernel — Function

JacksonKernel(n::Integer, N::Integer)

Jacksonkernel evaluated at n-th expansion coefficient with N in total (NC)

Lorentz Kernel is good for Green functions as it respects symmetry. The function LorentzKernels takes parameter λ and returns a kernel function.

KPM.LorentzKernels — Function

LorentzKernels(λ::Float64)

Returns function LorentzKernel(n, N) that evaluates Lorentz kernel with parameter λ, at n-th expansion coefficient with N in total (NC)

full API reference

KPM.JacksonKernel — Method

JacksonKernel(n::Integer, N::Integer)

Jacksonkernel evaluated at n-th expansion coefficient with N in total (NC)

KPM.LorentzKernels — Method

LorentzKernels(λ::Float64)

Returns function LorentzKernel(n, N) that evaluates Lorentz kernel with parameter λ, at n-th expansion coefficient with N in total (NC)

KPM.broadcast_dot_1d_1d! — Function

Vl and Vr are both [NH, NR] sized array. Each corresponding [:, NR] slice pair is dotted, saving into the target of [NR], multiplying by alpha and plus beta. Beta is either a number or vector of [NR]

KPM.broadcast_dot_2d_1d! — Method

Dot product each column of Vls with vector Vr, save in target.

target: 2D Array (n, NR), n >= ncols. Vls: 3D Array, shape (NH, NR, n), where n >= ncols. Vr: 2D Array, shape NH, NR ncols: Integer, number of columns.

KPM.broadcast_dot_reduce_avg_2d_1d! — Method

Dot product each column of Vls with vector Vr, save in target. Each view has NR replica of NH. This function take the average.

target: 1D Array (n), n >= ncols. Vls: 1D Array of 2D views, shape (n), each view (NH, NR), where n >= ncols. Vr: 2D Array, shape NH, NR ncols: Integer, number of columns.

KPM.chebyshev_iter — Method

evaluate from 3 to n ψall[:,3] comes from ψall[:,2] and ψall[:,1] separating into two function might improve performance (or not???)

KPM.dos0 — Function

dos0(μ, H_rescale_factor; NC, kernel, dE_order)

defined at /home/runner/work/KPM.jl/KPM.jl/src/applications/dos.jl:110.

Calculate DOS and its energy derivatives (by setting dE_order) at zero energy.

KPM.fermiFunction — Method

fermiFunction(E, E_f, beta)

calculate Fermi-Dirac function at energy E, Fermi energy μ and temperature β =1/T. Input and output all Float64. Infinite β only allowed when accessing fermi energy through fermiFunctions(). [For performance reason for now. TODO: allow β=Inf here withouth perf. reduction. ]

Allow sloppy use of type as long as convertion is available, if using keyword arguments.

KPM.fermiFunctions — Method

fermiFunctions(E_f::Float64, beta::Float64)

returns a fermi function with given E_f and beta.

Allow sloppy use of type as long as convertion is available, if using keyword arguments.

KPM.isNotBoundary — Method

give 0 for OBC=1 direction if i,i_ is on boundary. Otherwise 1

KPM.normalizeH — Method

Normalize H. If requested, allow renormalizing it to fixed value.

KPM.normalize_by_col — Method

Normalize a collection of vectors in an (NH, NR) array psi_in, where each column (that is psi_in[:, NRi]) represent a separate vector.

KPM.wrapAdd — Method

wrapAdd find the sum of x and y, with L+1=1